# service-base/src/main/webapp/statics/swagger/index.yaml
swagger: '2.0'
info:
  description: cofcoFrame是后台管理系统，采用流行的框架springMvc+spring+mybatis+shiro+redis+ehcache开发,实现了权限管理（菜单权限、数据权限），solr全文搜索引擎，activiti工作流程引擎，cas单点登陆等功能，完善的代码生成器 后期还会考虑改造成模块微服务化,做到模块的相对独立，使用更加灵活，努力做到快速开发OA办公系统
  version: 1.0.0
  title: cofcoFrame 接口文档
basePath: /frame-admin
schemes:
  - http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义接口数据
paths:
...
  /app/user/update:
    post:
      tags:
        - 用户管理
      summary: 修改用户
      produces:
        - application/json
      parameters:
        - name: body
          description: 用户对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/ApiUserEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /app/user/updatePassword:
    post:
      tags:
        - 用户管理
      summary: 修改当前用户密码
      produces:
        - application/json
      parameters:
        - name: passWord
          description: 原密码
          in: query
          type: string
          required: true
        - name: newPassWord
          description: 新密码
          in: query
          type: string
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: string
      msg:
        description: 消息
        type: string
...
  ApiUserEntity:
    type: object
    properties:
      id:
        description: 用户ID
        type: string
      userName:
        description: 用户名
        type: string
      loginName:
        description: 用户帐号
        type: string
      passWord:
        description: 密码
        type: string
      email:
        description: 邮箱
        type: string
      mobile:
        description: 手机号
        type: string
      myUpcomingCount:
        description: 待办条数
        type: int32
      myNoticeCount:
        description: 通知条数
        type: int32
      status:
        description: 状态  0：禁用   1：正常
        type: string
      createId:
        description: 创建者ID
        type: string
      createTime:
        description: 创建时间
        type: string
        format: date-time
...